+++ inject/inject-thrift-client/src/test/java/com/twitter/inject/thrift/integration/MultiJavaServerDarkTrafficFeatureTest.java
@@ -1,4 +1,4 @@
-package com.twitter.finatra.thrift.tests;
+package com.twitter.inject.thrift.integration;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -15,8 +15,8 @@ import org.junit.Test;
 
 import com.twitter.finagle.stats.InMemoryStatsReceiver;
 import com.twitter.finatra.thrift.EmbeddedThriftServer;
-import com.twitter.finatra.thrift.tests.doeverything.DoEverythingJavaThriftServer;
 import com.twitter.inject.server.PortUtils;
+import com.twitter.inject.thrift.integration.doeverything.DoEverythingJavaThriftServer;
 import com.twitter.util.Await;
 import com.twitter.util.Duration;
 import com.twitter.util.Future;
@@ -27,7 +27,7 @@ public class MultiJavaServerDarkTrafficFeatureTest extends Assert {
   private static final EmbeddedThriftServer DARK_THRIFT_SERVER =
       new EmbeddedThriftServer(
           new DoEverythingJavaThriftServer("dark-thrift-server"),
-          Collections.singletonMap("magicNum", "57"),
+          Collections.emptyMap(),
           Stage.DEVELOPMENT,
           true);
   private static final EmbeddedThriftServer LIVE_THRIFT_SERVER =
@@ -36,11 +36,11 @@ public class MultiJavaServerDarkTrafficFeatureTest extends Assert {
           Collections.unmodifiableMap(darkServerFlags()),
           Stage.DEVELOPMENT,
           true);
-  private static final com.twitter.doeverything.thriftjava.DoEverything.ServiceIface THRIFT_CLIENT =
+  private static final com.twitter.test.thriftjava.EchoService.ServiceIface THRIFT_CLIENT =
       LIVE_THRIFT_SERVER.thriftClient(
           THRIFT_CLIENT_ID,
           ClassTag$.MODULE$.apply(
-              com.twitter.doeverything.thriftjava.DoEverything.ServiceIface.class));
+              com.twitter.test.thriftjava.EchoService.ServiceIface.class));
 
   @AfterClass
   public static void tearDown() throws Exception {
@@ -56,7 +56,6 @@ public class MultiJavaServerDarkTrafficFeatureTest extends Assert {
 
   private static Map<String, String> darkServerFlags() {
     Map<String, String> flags = new HashMap<>();
-    flags.put("magicNum", "137");
     flags.put(
         "thrift.dark.service.dest",
         "/$/inet/" + PortUtils.loopbackAddress() + "/" + DARK_THRIFT_SERVER.thriftPort());
@@ -66,10 +65,10 @@ public class MultiJavaServerDarkTrafficFeatureTest extends Assert {
     return flags;
   }
 
-  /** magicNum is forwarded */
+  /** echo is forwarded */
   @Test
-  public void magicNum() throws Exception {
-    assertEquals("137", await(THRIFT_CLIENT.magicNum()));
+  public void echo() throws Exception {
+    assertEquals("137", await(THRIFT_CLIENT.echo("137")));
 
     // give a chance for the stat to be recorded on the live service
     LIVE_THRIFT_SERVER.assertHealthy(true);
@@ -91,8 +90,8 @@ public class MultiJavaServerDarkTrafficFeatureTest extends Assert {
 
   /** echo is forwarded */
   @Test
-  public void echo() throws Exception {
-    assertEquals("words", await(THRIFT_CLIENT.echo("words")));
+  public void setTimesToEcho() throws Exception {
+    assertEquals(Integer.valueOf(5), await(THRIFT_CLIENT.setTimesToEcho(5)));
 
     // give a chance for the stat to be recorded on the live service
     LIVE_THRIFT_SERVER.assertHealthy(true);