+++ inject/inject-thrift-client/src/test/java/com/twitter/inject/thrift/integration/doeverything/DoEverythingJavaThriftServer.java
@@ -1,8 +1,15 @@
-package com.twitter.finatra.thrift.tests.doeverything;
+package com.twitter.inject.thrift.integration.doeverything;
 
-import com.twitter.app.Flaggable;
+import java.util.Collection;
+
+import com.google.common.collect.ImmutableList;
+import com.google.inject.Module;
+
+import com.twitter.finagle.Filter;
+import com.twitter.finagle.Service;
 import com.twitter.finagle.ThriftMux;
 import com.twitter.finagle.tracing.NullTracer$;
+import com.twitter.finatra.annotations.DarkTrafficFilterType;
 import com.twitter.finatra.thrift.AbstractThriftServer;
 import com.twitter.finatra.thrift.filters.AccessLoggingFilter;
 import com.twitter.finatra.thrift.filters.ExceptionMappingFilter;
@@ -11,8 +18,6 @@ import com.twitter.finatra.thrift.filters.StatsFilter;
 import com.twitter.finatra.thrift.filters.ThriftMDCFilter;
 import com.twitter.finatra.thrift.filters.TraceIdMDCFilter;
 import com.twitter.finatra.thrift.routing.JavaThriftRouter;
-import com.twitter.finatra.thrift.tests.doeverything.exceptions.BarExceptionMapper;
-import com.twitter.finatra.thrift.tests.doeverything.exceptions.FooExceptionMapper;
 import com.twitter.util.NullMonitor$;
 
 public class DoEverythingJavaThriftServer extends AbstractThriftServer {
@@ -24,11 +29,11 @@ public class DoEverythingJavaThriftServer extends AbstractThriftServer {
 
     public DoEverythingJavaThriftServer(String name) {
         this.name = name;
-        flag().create(
-            "magicNum",
-            "26",
-            "Magic number",
-            Flaggable.ofString());
+    }
+
+    @Override
+    public Collection<Module> javaModules() {
+        return ImmutableList.<Module>of(new DoEverythingJavaDarkTrafficFilterModule());
     }
 
     @Override
@@ -41,6 +46,13 @@ public class DoEverythingJavaThriftServer extends AbstractThriftServer {
         return server
             .withMonitor(NullMonitor$.MODULE$)
             .withTracer(NullTracer$.MODULE$);
+    }
+
+    @Override
+    public Service<byte[], byte[]> configureService(Service<byte[], byte[]> service) {
+        return injector()
+            .instance(Filter.TypeAgnostic.class, DarkTrafficFilterType.class)
+            .andThen(service);
     }
 
     @Override
@@ -57,9 +69,6 @@ public class DoEverythingJavaThriftServer extends AbstractThriftServer {
             .filter(AccessLoggingFilter.class)
             .filter(StatsFilter.class)
             .filter(ExceptionMappingFilter.class)
-            .exceptionMapper(DoEverythingJavaExceptionMapper.class)
-            .exceptionMapper(FooExceptionMapper.class)
-            .exceptionMapper(BarExceptionMapper.class)
             .add(DoEverythingJavaThriftController.class);
     }